Removed install.sh.
It's not cross-platform. Also, I'm planning on packaging the JAR such that
all dependencies are included in the JAR. This doesn't allow easy sharing
of libraries, but, meh, Java doesn't seem to have a "standard" way of doing
that, anyway.

This used to be the setup of Snodes, and I kind of liked that.