Remove the snodes.command package
I never really understood the purpose of encapsulating every single
command as its own class. The more Java-esque way is to define action
listeners as anonymous classes or the like, so I redid the GUI code
with that in mind. It's still pretty clunky, and I think I could make
much better use of Java's inner classes features to reduce some
coupling, but it makes more sense than it did before; I found the
old architecture to be very difficult to read/follow.